1.不使用串口中断 2.更新flash读写接口 3.xmodem功能实现

+++ bootloader/Inc/user_flash.h
@@ -19,8 +19,8 @@
 #define USER_FLASH_DATA_END  0x0803FFFFU
 #define USER_FLASH_DATA_SIZE 0x00008000U /* 32k */
 
-extern void user_flash_read(uint16_t addr, uint16_t len, uint8_t *data);
-extern void user_flash_write(uint16_t addr, uint16_t len, uint8_t *data);
+extern void user_flash_read(uint32_t addr, uint16_t len, uint8_t *data);
+extern void user_flash_write(uint32_t addr, uint16_t len, uint8_t *data);
 extern void user_read_uuid(uint8_t *str);
 
 #endif /* __USER_FLASH_H */
